requirements.txt
-r requirements-dev.txt
aws-cdk-lib==2.50.0
constructs>=10.0.0,<11.0.0
Jinja2==3.1.2
rdk==0.8.5
rdklib==0.2.2
boto3==1.26.8
botocore==1.29.8
